move StyledLink component into Common folder
