+++ src/components/Homepage/blog.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { Row, Col } from '../grid'
 import { Padding } from 'styled-components-spacing'
 import { format } from 'date-fns'
-import StyledLink from '../styledLink'
+import StyledLink from '../Common/StyledLink'
 import { SectionTitle, Subtitle } from '../Typography'
 import Posts from '../posts'
 import CustomisedBulletpoint from '../CustomisedBulletpoint'
