+++ src/components/Homepage/jobs.js
@@ -4,7 +4,7 @@ import { Padding } from 'styled-components-spacing'
 import breakpoint from 'styled-components-breakpoint'
 import remcalc from 'remcalc'
 import styled from 'styled-components'
-import StyledLink from '../styledLink'
+import StyledLink from '../Common/StyledLink'
 import { SectionTitle, Subtitle } from '../Typography'
 import Jobs from '../jobsByLocation'
 import CustomisedBulletpoint from '../CustomisedBulletpoint'
