+++ src/components/OurWork/CaseStudy.js
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 import breakpoint from 'styled-components-breakpoint'
 import { Row, Col } from '../grid'
 import { Padding } from 'styled-components-spacing'
-import StyledLink from '../styledLink'
+import StyledLink from '../Common/StyledLink'
 import Image from '../Common/Image'
 import { CardTitle, BodyPrimary } from '../Typography'
 import getIntroSentence from '../../utils/getIntroSentence'
