+++ src/components/Speciality/blog.js
@@ -3,7 +3,7 @@ import { Padding } from 'styled-components-spacing'
 import { format } from 'date-fns'
 import { Row, Col, Grid } from '../grid'
 import { SectionTitle, Subtitle, BodyPrimary } from '../Typography'
-import StyledLink from '../styledLink'
+import StyledLink from '../Common/StyledLink'
 import Posts from '../posts'
 import CustomisedBulletpoint from '../CustomisedBulletpoint'
 import makeBlogPosts from '../../utils/makeBlogPosts'
