+++ src/components/Speciality/talks.js
@@ -4,7 +4,7 @@ import Flex from 'styled-flex-component'
 import { Row, Col, Grid } from '../grid'
 import { SectionTitle } from '../Typography'
 import { Padding } from 'styled-components-spacing'
-import StyledLink from '../styledLink'
+import StyledLink from '../Common/StyledLink'
 import BlueBackground from '../BlueBG'
 import styled from 'styled-components'
 import CompactVideoLink from '../Common/CompactVideoLink'
