+++ src/components/Speciality/training.js
@@ -3,7 +3,7 @@ import { Row, Col, Grid } from '../grid'
 import { SectionTitle, Subtitle, BodyPrimary } from '../Typography'
 import { Padding } from 'styled-components-spacing'
 import GreyBackground from '../GreyBG'
-import StyledLink from '../styledLink'
+import StyledLink from '../Common/StyledLink'
 
 const TrainingStage = ({ title, body, icon }) => (
   <Col width={[1, 1, 1, 1, 4 / 12]}>
