+++ src/components/Training/Modal.js
@@ -4,7 +4,7 @@ import breakpoint from 'styled-components-breakpoint'
 import is from 'styled-is'
 import ReactMarkdown from 'react-markdown'
 import { Padding } from 'styled-components-spacing'
-import StyledLink from '../styledLink'
+import StyledLink from '../Common/StyledLink'
 import Image from '../Common/Image'
 import remcalc from 'remcalc'
 import { SectionTitle, BodyPrimary } from '../Typography'
